Delete repo initialized .gitignore
I forgot that when you build a quarto site, initializing it will create a git ignore file. I don't want to create a conflict between the two, so I'm removing this file first. I copied the contents in a separate document for now if in case I need to add them to the quarto git ignore file